#25 | Update - 02.05.2026
Add:
- Change everyware INT and LONG to Double for better formatter huge score
- Fix a huges bugs with display formater main points
- New Event: Combo Chain
- New Prefab for Stats: Combo Chain
- New Skill: Auto Lucky BBonus Collector

+++ Assets/Scripts/GameManagement/Clicker/Clicker Prefabs/Prefab_Points.cs
@@ -5,11 +5,12 @@ public class Prefab_Points : MonoBehaviour
 {
     [SerializeField] TextMeshProUGUI totalPointsText;
 
-    public void UpdateTotalPointsPrefab(int amount)
+    public void UpdateTotalPointsPrefab(double amount)
     {
         if (totalPointsText == null) return;
 
         string formattedValue = NumberFormatter.FormatWithDots(amount);
+
         totalPointsText.text = $"Points \n{formattedValue}";
     }
 }
