+++ Assets/Scripts/GameManagement/Clicker/Clicker Stats/Clicker_Stats.cs
@@ -5,7 +5,7 @@ public class Clicker_Stats : MonoBehaviour
     [SerializeField] Total_Points totalPointsDisplay;
     [SerializeField] Total_PPS totalPPSDisplay;
 
-    public void UpdateAllStats(int totalPoints, int totalPPS)
+    public void UpdateAllStats(double totalPoints, double totalPPS)
     {
         if (totalPointsDisplay != null)
         {
